Move PremiumPill to common directory
