+++ assets/js/common/PremiumPill/PremiumPill.stories.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+import PremiumPill from '.';
+
+export default {
+  title: 'Components/PremiumPill',
+  component: PremiumPill,
+};
+
+export function Default() {
+  return <PremiumPill />;
+}
