+++ assets/js/common/PremiumPill/index.js
@@ -1,0 +1,3 @@
+import PremiumPill from './PremiumPill';
+
+export default PremiumPill;
