--- assets/js/components/PremiumPill/PremiumPill.stories.jsx
@@ -1,11 +1,0 @@
-import React from 'react';
-import PremiumPill from '.';
-
-export default {
-  title: 'Components/PremiumPill',
-  component: PremiumPill,
-};
-
-export function Default() {
-  return <PremiumPill />;
-}
