--- assets/js/components/PremiumPill/index.js
@@ -1,3 +1,0 @@
-import PremiumPill from './PremiumPill';
-
-export default PremiumPill;
